Adds support for the Z80 register sets
Adds support for the Z80 register sets based on the CPU Register section of the document starting on page 16.

The processor has the cool functionality of being able to switch between different sets of  registers on demand via a couple of CPU instructions. This has been added with the ability to switch between sets via the appropriate method call.